--- src/main/java/cn/thinkmoon/blog/modules/attachment/service/AttachmentService.java
@@ -1,18 +1,0 @@
-package cn.thinkmoon.blog.modules.attachment.service;
-
-import cn.thinkmoon.blog.core.property.QiniuProperty;
-import cn.thinkmoon.blog.utils.QiniuUtils;
-import org.springframework.stereotype.Service;
-
-import javax.annotation.Resource;
-
-@Service
-public class AttachmentService {
-
-    @Resource
-    private QiniuProperty qiniuProperty;
-
-    public String generateUploadCode(){
-        return QiniuUtils.generateUploadCode(qiniuProperty);
-    }
-}
