--- src/main/java/cn/thinkmoon/blog/modules/content/dao/UserDAO.java
@@ -1,8 +1,0 @@
-package cn.thinkmoon.blog.modules.content.dao;
-
-import cn.thinkmoon.blog.modules.content.pojo.po.UserPO;
-import com.baomidou.mybatisplus.core.mapper.BaseMapper;
-
-public interface UserDAO extends BaseMapper<UserPO> {
-    UserPO queryUserByAccount(String account);
-}
