--- src/main/java/cn/thinkmoon/blog/modules/content/pojo/po/CategoryPO.java
@@ -1,18 +1,0 @@
-package cn.thinkmoon.blog.modules.content.pojo.po;
-
-import com.baomidou.mybatisplus.annotation.IdType;
-import com.baomidou.mybatisplus.annotation.TableId;
-import com.baomidou.mybatisplus.extension.activerecord.Model;
-import lombok.Data;
-
-@Data
-public class CategoryPO extends Model<CategoryPO> {
-    private static final long serialVersionUID = 1L;
-
-    @TableId(value = "mid", type = IdType.AUTO)
-    private Integer mid;
-
-    private String name;
-
-    private Integer count;
-}
