--- src/main/java/cn/thinkmoon/blog/modules/content/service/CategoryService.java
@@ -1,10 +1,0 @@
-package cn.thinkmoon.blog.modules.content.service;
-
-import cn.thinkmoon.blog.modules.content.pojo.po.CategoryPO;
-
-import java.util.List;
-
-public interface CategoryService {
-
-    List<CategoryPO> queryList(String name);
-}
