--- src/main/java/cn/thinkmoon/blog/modules/content/service/EmailService.java
@@ -1,41 +1,0 @@
-package cn.thinkmoon.blog.modules.content.service;
-
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.mail.javamail.JavaMailSender;
-import org.springframework.mail.javamail.MimeMessageHelper;
-import org.springframework.stereotype.Component;
-
-import javax.mail.MessagingException;
-import javax.mail.internet.MimeMessage;
-
-/**
- * The Email service.
- *
- * @author felord.cn
- * @since 2020 /1/14 23:22
- */
-@Component
-public class EmailService {
-    @Autowired
-    private JavaMailSender javaMailSender;
-    @Value("${spring.mail.username}")
-    private String from;
-
-    /**
-     * 发送纯文本邮件.
-     * @param to      目标email 地址
-     * @param subject 邮件主题
-     * @param text    纯文本内容
-     */
-    public void sendMail(String to, String subject, String text) throws MessagingException {
-        MimeMessage mimeMessage = javaMailSender.createMimeMessage();
-        MimeMessageHelper helper=new MimeMessageHelper(mimeMessage,true);
-        helper.setFrom(from);
-        helper.setTo(to);
-        helper.setSubject(subject);
-
-        helper.setText(text,true);
-        javaMailSender.send(mimeMessage);
-    }
-}
