--- src/main/java/cn/thinkmoon/blog/modules/content/service/UserService.java
@@ -1,17 +1,0 @@
-package cn.thinkmoon.blog.modules.content.service;
-
-import cn.thinkmoon.blog.modules.content.pojo.po.UserPO;
-import cn.thinkmoon.blog.modules.content.dao.UserDAO;
-import com.baomidou.mybatisplus.extension.service.impl.ServiceImpl;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Service;
-
-@Service
-public class UserService extends ServiceImpl<UserDAO, UserPO> {
-    @Autowired
-    private UserDAO userMapper;
-
-    public UserPO queryUserByAccount(String account) {
-        return userMapper.queryUserByAccount(account);
-    }
-}
