+++ src/main/java/cn/thinkmoon/blog/modules/dao/UserDAO.java
@@ -1,0 +1,8 @@
+package cn.thinkmoon.blog.modules.dao;
+
+import cn.thinkmoon.blog.modules.pojo.UserPO;
+import com.baomidou.mybatisplus.core.mapper.BaseMapper;
+
+public interface UserDAO extends BaseMapper<UserPO> {
+    UserPO queryUserByAccount(String account);
+}
