+++ src/main/java/cn/thinkmoon/blog/modules/pojo/FieldsPO.java
@@ -1,0 +1,12 @@
+package cn.thinkmoon.blog.modules.pojo;
+
+import lombok.Data;
+
+@Data
+public class FieldsPO {
+    private int cid;
+
+    private String name;
+
+    private String value;
+}
