+++ src/main/java/cn/thinkmoon/blog/modules/pojo/UserPO.java
@@ -1,0 +1,22 @@
+package cn.thinkmoon.blog.modules.pojo;
+
+import com.baomidou.mybatisplus.annotation.IdType;
+import com.baomidou.mybatisplus.annotation.TableId;
+import com.baomidou.mybatisplus.extension.activerecord.Model;
+import com.fasterxml.jackson.annotation.JsonIgnore;
+import lombok.Data;
+
+@Data
+public class UserPO extends Model<UserPO> {
+    private static final long serialVersionUID = 1L;
+
+    @TableId(value = "id", type = IdType.AUTO)
+    private Integer id;
+
+    private String name;
+
+    private String account;
+
+    @JsonIgnore
+    private String password;
+}
