--- src/main/java/cn/thinkmoon/blog/modules/post/pojo/FieldsPO.java
@@ -1,12 +1,0 @@
-package cn.thinkmoon.blog.modules.post.pojo;
-
-import lombok.Data;
-
-@Data
-public class FieldsPO {
-    private int cid;
-
-    private String name;
-
-    private String value;
-}
